% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_equivalents}
\alias{add_equivalents}
\title{Get Malaise samples and Microgastrinae tubes linked to a wasp jar}
\usage{
add_equivalents(x)
}
\arguments{
\item{x}{Data frame with at least columns "wasp_jar" and "no_microgastrinae". Usually exported from file "Samples sorted.numbers".}
}
\value{
The data frame (\code{x}) with added columns "sample", "microgastrinae_tube" and "microgastrinae". "microgastrinae" is blank if no Microgastrinae were found in the sample.
}
\description{
This is a helper function used by \code{\link[=upload_wasps]{upload_wasps()}} and \code{\link[=update_sorted]{update_sorted()}}. For each given wasp jar,   gets the identifier of the Malaise sample it came from, and of the Microgastrinae tube that came from it.
}
\seealso{
\code{\link[=upload_wasps]{upload_wasps()}} and \code{\link[=update_sorted]{update_sorted()}}, which use this function.
}
