% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_package_and_data.R
\docType{data}
\name{equivalents}
\alias{equivalents}
\title{Ugandan wasp jars and their matching Malaise samples and Microgastrinae tubes}
\format{
A data frame with 876 rows (one row for each sample) and 6 variables. The variables are:
\itemize{
\item wasp_jar Identifier of a wasp jar. Short form, e.g. "6919".
\item microgastrinae_tube Identifier of a microgastrinae tube. Short form, e.g. "4770"
\item sample Identifier of a Malaise sample. Short form, e.g. "CCT1-141022".
}

...and three columns with the same identifiers in long format, e.g. "http://mus.utu.fi/ZMUT.6919", "http://mus.utu.fi/ZMUT.4470", "http://mus.utu.fi/ZMUT.CCT1-141022".
}
\usage{
equivalents
}
\description{
Identifiers of Ugandan wasp jars, with the Malaise samples they came from and the Microgastrinae tubes that were (or will be) split from them.
}
\keyword{datasets}
