% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_events.R
\name{update_events}
\alias{update_events}
\title{Upload wasp jar data (events) in Kotka}
\usage{
update_events(
  x,
  f = "uploads/new_events.csv",
  removed = c("Banchinae", "Microgastrinae", "Ophioninae", "Pimplinae",
    "Tryphoninae:Netelia")
)
}
\arguments{
\item{x}{Data frame with at least column "wasp_jar". Usually exported from file "Samples sorted.numbers".}

\item{f}{File path where to store the upload file. Will refuse to overwrite existing files.}

\item{removed}{Character vector of what taxa have been removed.}
}
\value{
If the file already exists or the folder can't be found, returns a message.
}
\description{
Creates Kotka events to mark a jar as having been checked (e.g. "2023-01-27 Pimplinae removed"). Also creates a second file that lists the wasp jar IDs, so they are easy to search for in Kotka.
}
