% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_wasps.R
\name{upload_wasps}
\alias{upload_wasps}
\title{Upload pinned wasps to Kotka}
\usage{
upload_wasps(x, f = "uploads/kotka_upload_wasps.csv")
}
\arguments{
\item{x}{Data frame with at least columns "wasp_jar", "nwasps" and "no_microgastrinae". Usually exported from file "Samples sorted.numbers".}

\item{f}{File path where to store the upload file. Will refuse to overwrite existing files.}
}
\value{
If the file already exists or the folder can't be found, returns a message.
}
\description{
Creates a Kotka upload file for pinned wasps. Everything in the file should be ready except for the specimen IDs (but worth looking through the file just in case).
}
